refactor: allow dependency injection of httpClient
Previously the GuzzleHttp Client was created inside the Request helper which it self was instanciated in the Client. This made it impossible to dependency inject a http Client of my own choosing.

While making DI possible I opted for removing the Request class entirely as it wasn’t creating much value.